[postgres] Move fields.go and to sources/postgres/transformer
