+++ sources/postgres/transformer/fields.go
@@ -1,4 +1,4 @@
-package debezium
+package transformer
 
 import (
 	"github.com/artie-labs/transfer/lib/debezium"
