+++ sources/postgres/transformer/fields_test.go
@@ -1,4 +1,4 @@
-package debezium
+package transformer
 
 import (
 	"testing"
